Computer Vision for image classifications of rosted coffe bean/deployment/requirements.txt
streamlit
tensorflow-cpu
pandas
numpy
